+++ src/node.rs
@@ -15,6 +15,7 @@ pub use iters::{
 };
 pub use node_data::{Element, NodeData};
 pub use node_ref::{Node, NodeRef};
+pub(crate) use serializing::format_text;
 pub use serializing::SerializableNodeRef;
 
 /// Represents a Node ID.