+++ tests/data.rs
@@ -104,3 +104,28 @@ pub static ATTRS_CONTENTS: &str = r#"<!DOCTYPE html>
             </div>
         </body>
     </html>"#;
+
+pub static DMC_CONTENTS: &str = r#"<!DOCTYPE html>
+<html>
+    <head></head>
+    <body>
+        <div id="main">
+            <div>
+                <p>Listen up y'all, it's time to get down<br>
+                'Bout that <b>normalized_char_count</b> in this town<br>
+                Traversing nodes with style and grace<br>
+                Counting chars at a steady pace</p>
+            </div>
+
+            <div>
+                <p>No split whitespace, that's old school<br>
+                Direct counting's our golden rule<br>
+                Skip them nodes that ain't text or element<br>
+                That's how we keep our code development!</p>
+            </div>
+            <pre>
+            WORD!
+            </pre>
+        </div>
+    </body>
+</html>"#;